Proposed wording for high-entrropy inputs, gender neutrality
Wording:

High-quality entropy source is a first swing at an improvement.

We could lean into the NIST DRBG definitions if we wanted precise terminology, but I feel there's often a contention between precision and clarity, and security consideration documents scream "focus on clarity" to me.  I'm open to dissenting opinions here, of course.

Pronouns:

The adversary is not a specific actor and thus neutral pronouns (they'them) are more suitable. If a specific actor was in mind (e.g., Mallory), rather than just a vague "adversary", then this would be an incorrect change.